Fix double typo in LZMAStream#initialize. Fixes #6.
This was a curious combination of two typos; one missing NOT followed
by an invalid field name (missing the "_in"). Due to the missing NOT,
the missing field was never to be encountered, and the `super' call
was executed instead, which errored on JRuby due to incorrect arguments.
The missing NOT caused the underlying `lzma_stream' struct to be
created uninitialized, which just worked *by chance* (as it always is
if you use variables just created on unused memory in C). Now it properly
gets set up as liblzma would do it itself.